requirements/development.txt
-r base.txt
factory-boy==2.12.0
expects==0.9.0
django-nose==1.4.6
coverage==4.5.4
rednose==1.3.0
